requirements.txt
langchain
langchain-google-genai
langchain-community
langgraph
python-dotenv
duckduckgo-search
numexpr
ddgs
